cliente_funcionario_server/requirements.txt
passlib==1.7.4
psycopg2-binary==2.9.10
PyQt5==5.13.2
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
python-dotenv==1.0.1
pandas==2.2.3 